requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-appconf==1.0.5
django-compressor==4.4
mysqlclient==2.2.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
typing-extensions==4.8.0
whitenoise==6.6.0
